pkg/endpoint: move policySignal type definition, cleanPolicySignals to policyrevision.go
Signed-off by: Ian Vernon <ian@cilium.io>
